+++ src/test/java/wooteco/subway/path/domain/FareCalculatorTest.java
@@ -13,7 +13,6 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 import wooteco.subway.line.domain.Line;
-import wooteco.subway.path.application.FareCalculator;
 
 class FareCalculatorTest {
     private FareCalculator fareCalculator;
